+++ src/test/run-pass/task-spawn-move-and-copy.rs
@@ -19,10 +19,10 @@ pub fn main() {
     let (tx, rx) = channel::<usize>();
 
     let x: Box<isize> = box 1;
-    let x_in_parent = &(*x) as *const isize;
+    let x_in_parent = &(*x) as *const isize as usize;
 
     let t = thread::spawn(move || {
-        let x_in_child = &(*x) as *const isize;
+        let x_in_child = &(*x) as *const isize as usize;
         tx.send(x_in_child).unwrap();
     });
